Block proto sid-->long

+++ hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/protocol/ExtendedBlock.java
@@ -52,7 +52,7 @@ public class ExtendedBlock {
     block = new Block(blkid, len, genstamp);
   }
 
-  public ExtendedBlock(final String poolId, final long blkid, final int sid, final long len,
+  public ExtendedBlock(final String poolId, final long blkid, final long sid, final long len,
       final long genstamp) {
     this.poolId = poolId;
     block = new Block(blkid, sid, len, genstamp);
